+++ app/[locale]/layout.tsx
@@ -1,17 +1,7 @@
-import { locales, type Locale } from '@/constants';
+import type { Locale } from '@/constants';
 import { Providers } from '@/features/providers';
 
-import type { Metadata } from 'next';
 import type { ReactNode } from 'react';
-
-export const metadata: Metadata = {
-  title: 'Pokemons',
-  description: 'Description of pokemons',
-};
-
-export function generateStaticParams() {
-  return locales.map((locale) => ({ locale }));
-}
 
 export default async function RootLayout({
   children,
